moved the big formobject in the story to separate file
